cutedsl/requirements.txt
# This library requires Python >= 3.12
nvidia-cutlass-dsl==4.1.0.dev0
torch